local ok, server not ok

+++ services/agency_service/src/main/java/soa/study/agency_service/rest/dto/FlatStatResponse.java
@@ -16,7 +16,7 @@ public class FlatStatResponse {
 
     private Long flatId;
 
-    private Integer numberOfRooms;
+    private Long numberOfRooms;
 
     private Float price;
